requirements.txt
python-dotenv
requests
edge-tts
google-genai
ffmpeg-python
pytest
pillow
pydub
